Update InlineEditBox.js
I have committed the lines
this.editable.innerHTML = this.value;

to

this.editable.textContent= this.value;

and 

this.editable.innerHTML = tt;

to 

this.editable.textContent = tt;


This will resolve the vulnerability because textContent is a safe alternative because it only sets plain text. This will not allow HTML, tags, scripts, attributes because they will not be executed or rendered.

All the scripts will be converted to PLAINTEXT and not as javascript.